refactor: Eliminate code duplication between index.ts and test-index.ts
- Created shared ipc-handlers.ts module for IPC handlers
- Created shared menu.ts module for menu creation
- Both index.ts and test-index.ts now use the same shared code
- Removed deterministic if-check in tests as requested
- All tests still passing
